Rename application file
