+++ cdk/kraken/src/index.ts
@@ -1,10 +1,10 @@
-export * from './application';
 export * from './cdk';
 export * from './deploy';
 export * from './django-project';
 export * from './django';
 export * from './docker';
 export * from './integration-tests';
+export * from './labs-application';
 export * from './pypi';
 export * from './react-project';
 export * from './react';
